Check that the test name is a symbol before registering the test.
Since tests are set on the test property of the test name (i.e. with
GET), if the name is not a symbol the registration will fail. But
since the prior step will have registered the test with the package
suite, whenever the suite for that package is run later it will signal
an error. This way, the error prevents the bogus tests from being
registered and corrupting the package suite.