Fix calculation of test duration.
Test duration is in milliseconds, but it is calculated using internal
time units. On implementations where INTERNAL-TIME-UNITS-PER-SECOND is
not 1000, this will produce wildly inflated test duration values in
the reports. This fixes the calculation to account for varying time
resolution in implementations so that test duration is always in
milliseconds.